pyproject.toml
[project]
name = "rag-project2"
version = "0.1.0"
description = "Production-ready RAG application with multi-chat support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.4.0",
    "fastapi>=0.128.0",
    "langchain>=1.2.3",
    "langchain-chroma>=1.1.0",
    "langchain-cohere>=0.5.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langchain-postgres>=0.0.16",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.6",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.21",
    "qdrant-client>=1.16.2",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "unstructured[pdf]>=0.18.27",
    "uvicorn>=0.40.0",
    "watchfiles[watchdog]>=1.1.1",
]
